+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Add `backgroundColor` props on `CrowdfundingCard` component.
+- Add `backgroundColor` and `color` props on `CrowdfundingCard` component.
 
 ## [20.4.0] - 2018-03-26